généricité

+++ TextEditor/tests/test2.cpp
@@ -19,7 +19,7 @@ void test_case() {
     unsigned long long end = editor.getSelectionEnd();
 
     Tools::assert_equals(start,end, "start et end ne sont pas égaux à l'init");
-    Tools::assert_equals(start ,14, "start n'est pas à la fin de la string à l'init");
+    Tools::assert_equals(start ,(unsigned long long)14, "start n'est pas à la fin de la string à l'init");
 
     editor.keyPressed(false,false,false,SDLK_RIGHT);
